Refactor, from downloadClangDoc to downloadTool
